Performance improvements with more async algorithm

+++ src/sensor/config.rs
@@ -4,7 +4,7 @@ use serde::Deserialize;
 pub struct SensorConfig {
     pub name: String,
     pub pin_number: u8,
-    pub refresh_rate_milisecs: u64,       // miliseconds
+    pub refresh_rate_milisecs: u128,      // miliseconds
     pub motion_time_period_milisecs: u64, // miliseconds
     pub minimal_triggering_number: i16,
 }
